pyproject.toml
[project]
name = "hvac-stability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "environ-config>=24.1.0",
    "pykumo>=0.3.11",
    "typer>=0.16.0",
]

[project.scripts]
hvac-stability = "hvac_stability.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hvac_stability"]
